dom/tests/mochitest/pointerlock/file_loseFocusWindow.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=633602
-->
<head>
  <title>Bug 633602 - file_loseFocusWindow.html</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js">
  </script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js">
  </script>
  <script type="application/javascript" src="pointerlock_utils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
  <a target="_blank"
    href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">
    Mozilla Bug 633602
  </a>
  <div id="div"></div>
  <pre id="test">
    <script type="text/javascript">
      /*
       * Test for Bug 633602
       * If element has the pointer locked and window loses focus
       * pointer should be unlocked
       */

      SimpleTest.waitForExplicitFinish();

      var div
        , newWindow = null
        , loseFocusUnlock = false;

      function runTests () {
        ok(loseFocusUnlock, "Pointer should be unlocked if window loses focus");
      }

      document.addEventListener("mozpointerlockchange", function (e) {
        if (document.mozPointerLockElement === div) {
          divPointerLocked = true;
          // Open new window to blur current one
          newWindow = window.open("file_loseFocusWindow-child.html", "", "width=500,height=500");
        } else {
          loseFocusUnlock = true;
        }
      }, false);

      function finish() {
        runTests();
        SimpleTest.finish();
      }

      // Will get triggered by newWindow when it's open and focused.
      function childReady() {
        newWindow.close();
        // On OS X Lion, tests cause problems. Timeouts are a bad way to get around
        // the problem and may lead to future [orange], but they are the only option
        // at this point.
        SimpleTest.waitForFocus(function() { setTimeout(finish, 3000); });
      }

      document.addEventListener("mozfullscreenchange", function() {
        if (document.mozFullScreenElement === div) {
          div.mozRequestPointerLock();
        }
      }, false);

      function start() {
        div = document.getElementById("div");
        div.mozRequestFullScreen();
      }
    </script>
  </pre>
</body>
</html>
